requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
anthropic>=0.40.0
Pillow>=11.0.0
pydantic-settings>=2.6.0
python-multipart>=0.0.17

# Phase 4 (uncomment when ready):
# open-clip-torch>=2.26.1
# sqlite-vec>=0.1.0
# torch>=2.0.0
